generate music from a .smus file, still not used anywhere yet

+++ Source/GameState/EntityState.h
@@ -87,7 +87,7 @@ public:
 class DynamicCameraAnchor: public EntityState {
 private:
 	static constexpr float speedPerSecond = 120.0f;
-	static constexpr float diagonalSpeedPerSecond = speedPerSecond * MathUtils::sqrtConst(0.5f);
+	static constexpr float diagonalSpeedPerSecond = speedPerSecond * (float)MathUtils::sqrtConst(0.5);
 	static constexpr float sprintModifier = 2.0f;
 
 	ReferenceCounterHolder<DynamicValue> screenOverlayR;
